consumer/requirements.txt
kafka-python
rfc3339
iso8601
prometheus-client
